+++ packages/turbo-stream/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## master
 
+- Add `connected` state to stream source element. ([@palkan][])
+
+Backport of [turbo-rails#430](https://github.com/hotwired/turbo-rails/pull/430).
+
+## 0.3.0 (2023-02-06)
+
+- Upgrade to `@anycable/web` v0.6.0.
+
 ## 0.1.0 (2022-07-14)
 
 - Major refactoring: use channel classes (allow custom), ignore Turbo cached pages, avoid waiting for subscribe (to unsubscribe on element disconnect). ([@palkan][])